fix: 修复7个代码Bug和死代码问题
- fix: Long类型转换使用Long.parseLong替代Integer.parseInt
- fix: dateDiff参数顺序错误导致startDate被忽略
- fix: driverMajorVersion获取了错误的DatabaseMetaData方法
- fix: concat_ws抛出UnsupportedOperationException
- fix: buildDrop方法中DbType.MySQL死代码分支改为DbType.SQLServer
- fix: buildCopy方法中死代码分支修复
- fix: SqlBeanMapper空catch块改为记录日志

+++ flexsql-core/src/main/java/cn/vonce/sql/uitls/SqlBeanUtil.java
@@ -1423,7 +1423,7 @@ public class SqlBeanUtil {
             return Integer.parseInt(value.toString());
         }
         if (type == Long.class || type == long.class) {
-            return Integer.parseInt(value.toString());
+            return Long.parseLong(value.toString());
         }
         if (type == Float.class || type == float.class) {
             return Float.parseFloat(value.toString());
